pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pokemon-superhuman"
version = "0.1.0"
description = "Superhuman Pokemon AI via Scaled Offline RL and Self-Play"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Pokemon AI Project"}
]
keywords = ["pokemon", "reinforcement-learning", "transformer", "self-play"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "accelerate>=0.25.0",
    "einops>=0.7.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "wandb>=0.16.0",
    "pyyaml>=6.0.0",
    "tqdm>=4.66.0",
    "tensorboard>=2.15.0",
    "safetensors>=0.4.0",
    "aiohttp>=3.9.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "black>=23.12.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
]
distributed = [
    "deepspeed>=0.12.0",
]
h100 = [
    "deepspeed>=0.12.0",
    "flash-attn>=2.4.0",
]
full = [
    "deepspeed>=0.12.0",
    "flash-attn>=2.4.0",
    "poke-env>=0.7.0",
    "gymnasium>=0.29.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "datasets>=2.15.0",
    "huggingface-hub>=0.19.0",
    "redis>=5.0.0",
    "ray[default]>=2.9.0",
    "websockets>=12.0",
    "orjson>=3.9.0",
    "typer>=0.9.0",
]

[project.scripts]
pokemon-train = "pokemon_ai.train:main"
pokemon-eval = "pokemon_ai.evaluate:main"
pokemon-play = "pokemon_ai.play:main"
pokemon-league = "pokemon_ai.league:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
